refactor: #102 - lib auth  파일을 feature별 경로로 정리
