+++ src/app/api/auth/resend-verification-email/route.test.ts
@@ -1,11 +1,11 @@
 import { NextRequest } from "next/server";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { resetResendRateLimitStore } from "@/lib/auth/checkResendRateLimit";
-import { getLastVerificationResendAt } from "@/lib/auth/getLastVerificationResendAt";
-import { resendVerificationEmail } from "@/lib/auth/resendVerificationEmail";
-import { setLastVerificationResendAt } from "@/lib/auth/setLastVerificationResendAt";
-import { AUTH_API_CODES } from "@/lib/constants/authApiCodes";
+import { AUTH_API_CODES } from "@/features/auth/constants/authApiCodes";
+import { resetResendRateLimitStore } from "@/features/auth/resend-verification-email/lib/checkResendRateLimit";
+import { getLastVerificationResendAt } from "@/features/auth/resend-verification-email/lib/getLastVerificationResendAt";
+import { resendVerificationEmail } from "@/features/auth/resend-verification-email/lib/resendVerificationEmail";
+import { setLastVerificationResendAt } from "@/features/auth/resend-verification-email/lib/setLastVerificationResendAt";
 
 import { POST } from "./route";
 vi.mock("@/lib/auth/getLastVerificationResendAt");
