+++ src/features/auth/lib/getUserByEmail.ts
